+++ ubo_app/services/080-docker/reducer.py
@@ -168,14 +168,22 @@ IMAGES = {
                     r'^[a-zA-Z0-9]{20,30}_[a-zA-Z0-9]{20,30}$',
                     resolver=lambda code, _: code,
                     prompt='Enter the Ngrok Auth Token',
-                    extra_information="""Get the token from {ngrok|EH N G EH R AA K} \
-dashboard, convert it to {QR|K Y UW AA R} code and scan it here.""",
+                    extra_information="""\
+Follow these steps:
+
+1. Login to your {ngrok|EH N G EH R AA K} account
+2. Get the authentication token from the dashboard
+3. Convert it to {QR|K Y UW AA R} code
+4. Scan QR code to input the token""",
                 ),
             },
             command=lambda: qrcode_input(
                 '',
                 resolver=lambda code, _: code,
                 prompt='Enter the command, for example: `http 80` or `tcp 22`',
+                extra_information="""\
+This is the command you would enter when running {ngrok|EH N G EH R AA K}.
+Refer to {ngrok|EH N G EH R AA K} documentation for further information""",
             ),
         ),
         *(